requirements.txt
torch>=2.0.0
gymnasium>=0.28.0
ale-py>=0.8.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
opencv-python>=4.8.0
pillow>=10.0.0
tqdm>=4.65.0
wandb>=0.15.0
scikit-learn>=1.3.0
pandas>=2.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
pytest>=7.4.0
pytest-cov>=4.1.0
pyyaml>=6.0.0
